--- a/src/ExtendedXmlSerializer/ContentModel/AliasesBase.cs
+++ b/src/ExtendedXmlSerializer/ContentModel/AliasesBase.cs
@@ -26,7 +26,7 @@ using ExtendedXmlSerialization.Core.Sources;
 
 namespace ExtendedXmlSerialization.ContentModel
 {
-	public abstract class AliasesBase<T> : WeakCacheBase<T, string>, IAliases<T> where T : MemberInfo
+	public abstract class AliasesBase<T> : ReferenceCacheBase<T, string>, IAliases<T> where T : MemberInfo
 	{
 		// public abstract string Get(T parameter);
 	}