--- a/src/ExtendedXmlSerializer/ContentModel/Content/DecoratedContentOption.cs
+++ b/src/ExtendedXmlSerializer/ContentModel/Content/DecoratedContentOption.cs
@@ -1,6 +1,6 @@
-﻿// MIT License
+// MIT License
 // 
-// Copyright (c) 2016 Wojciech Nagórski
+// Copyright (c) 2016 Wojciech Nag�rski
 //                    Michael DeMond
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -21,19 +21,13 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
+using System.Reflection;
+using ExtendedXmlSerialization.Core.Sources;
 
-namespace ExtendedXmlSerialization.Core.Sources
+namespace ExtendedXmlSerialization.ContentModel.Content
 {
-	public class DelegatedSource<TParameter, TResult> : IParameterizedSource<TParameter, TResult>
+	class DecoratedContentOption : DecoratedOption<TypeInfo, ISerializer>, IContentOption
 	{
-		readonly Func<TParameter, TResult> _source;
-
-		public DelegatedSource(Func<TParameter, TResult> source)
-		{
-			_source = source;
-		}
-
-		public virtual TResult Get(TParameter parameter) => _source(parameter);
+		public DecoratedContentOption(IContentOption source) : base(source, source) {}
 	}
 }