--- a/src/ExtendedXmlSerializer/ContentModel/Converters/CachedConverter.cs
+++ b/src/ExtendedXmlSerializer/ContentModel/Converters/CachedConverter.cs
@@ -1,6 +1,6 @@
-﻿// MIT License
+// MIT License
 // 
-// Copyright (c) 2016 Wojciech Nagórski
+// Copyright (c) 2016 Wojciech Nag�rski
 //                    Michael DeMond
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -21,24 +21,26 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+using System;
+using System.Reflection;
+using ExtendedXmlSerialization.Core.Sources;
 using ExtendedXmlSerialization.Core.Specifications;
 
-namespace ExtendedXmlSerialization.Core.Sources
+namespace ExtendedXmlSerialization.ContentModel.Converters
 {
-	public abstract class CompositeOptionBase<TParameter, TResult> : Selector<TParameter, TResult>,
-	                                                                 IOption<TParameter, TResult>
+	class OptimizedConverterAlteration : IAlteration<IConverter>
 	{
-		readonly ISpecification<TParameter> _specification;
+		public static OptimizedConverterAlteration Default { get; } = new OptimizedConverterAlteration();
+		OptimizedConverterAlteration() {}
 
-		protected CompositeOptionBase(params IOption<TParameter, TResult>[] options)
-			: this(new AnySpecification<TParameter>(options), options) {}
+		public IConverter Get(IConverter parameter) => new CachedConverter(parameter);
+	}
 
-		protected CompositeOptionBase(ISpecification<TParameter> specification, params IOption<TParameter, TResult>[] options)
-			: base(options)
-		{
-			_specification = specification;
-		}
+	class CachedConverter : Converter<object>
+	{
+		public CachedConverter(IConverter converter) : this(converter, converter.Parse, converter.Format) {}
 
-		public bool IsSatisfiedBy(TParameter parameter) => _specification.IsSatisfiedBy(parameter);
+		public CachedConverter(ISpecification<TypeInfo> specification, Func<string, object> deserialize, Func<object, string> serialize)
+			: base(specification, new Cache<string, object>(deserialize).Get, new Cache<object, string>(serialize).Get) {}
 	}
 }