--- a/src/ExtendedXmlSerializer/ContentModel/Converters/IConverter.cs
+++ b/src/ExtendedXmlSerializer/ContentModel/Converters/IConverter.cs
@@ -34,11 +34,4 @@ namespace ExtendedXmlSerialization.ContentModel.Converters
 
 		string Format(T instance);
 	}
-
-	/*public interface IConverter : ISpecification<TypeInfo>
-	{
-		object Load(string data);
-
-		string Save(object instance);
-	}*/
 }