--- a/src/ExtendedXmlSerializer/ContentModel/Members/MemberInformation.cs
+++ b/src/ExtendedXmlSerializer/ContentModel/Members/MemberInformation.cs
@@ -25,17 +25,10 @@ using System.Reflection;
 
 namespace ExtendedXmlSerialization.ContentModel.Members
 {
-	public struct MemberInformation // : IEquatable<MemberInformation>
+	public struct MemberInformation
 	{
-		/*
-		readonly int _code;
-
 		public MemberInformation(MemberInfo metadata, TypeInfo memberType, bool assignable)
-			: this(metadata, memberType, assignable, metadata.MetadataToken.GetHashCode()) {}*/
-
-		public MemberInformation(MemberInfo metadata, TypeInfo memberType, bool assignable/*, int code*/)
 		{
-			/*_code = code;*/
 			Metadata = metadata;
 			MemberType = memberType;
 			Assignable = assignable;
@@ -44,16 +37,5 @@ namespace ExtendedXmlSerialization.ContentModel.Members
 		public MemberInfo Metadata { get; }
 		public TypeInfo MemberType { get; }
 		public bool Assignable { get; }
-
-		/*public bool Equals(MemberInformation other) => _code == other._code;
-
-		public override bool Equals(object obj)
-			=> !ReferenceEquals(null, obj) && obj is MemberInformation && Equals((MemberInformation) obj);
-
-		public override int GetHashCode() => _code;
-
-		public static bool operator ==(MemberInformation left, MemberInformation right) => left.Equals(right);
-
-		public static bool operator !=(MemberInformation left, MemberInformation right) => !left.Equals(right);*/
 	}
 }