--- a/src/ExtendedXmlSerializer/ContentModel/Xml/Namespacing/OptimizedXmlWriterFactory.cs
+++ b/src/ExtendedXmlSerializer/ContentModel/Xml/Namespacing/OptimizedXmlWriterFactory.cs
@@ -31,7 +31,7 @@ namespace ExtendedXmlSerialization.ContentModel.Xml.Namespacing
 		readonly IXmlWriterFactory _factory;
 		readonly IObjectNamespaces _namespaces;
 
-		public OptimizedXmlWriterFactory() : this(new Containers()) {}
+		public OptimizedXmlWriterFactory() : this(Containers.Default) {}
 
 		public OptimizedXmlWriterFactory(IContainers containers)
 			: this(XmlWriterFactory.Default, new ObjectNamespaces(new Members.Members(new Selector(containers)))) {}