--- a/src/ExtendedXmlSerializer/ContentModel/Xml/ObjectTypeWalker.cs
+++ b/src/ExtendedXmlSerializer/ContentModel/Xml/ObjectTypeWalker.cs
@@ -31,7 +31,7 @@ using ExtendedXmlSerialization.Core.Sources;
 
 namespace ExtendedXmlSerialization.ContentModel.Xml
 {
-	public class ObjectTypeWalker : ObjectWalkerBase<object, IEnumerable<TypeInfo>>, ISource<IEnumerable<TypeInfo>>
+	class ObjectTypeWalker : ObjectWalkerBase<object, IEnumerable<TypeInfo>>, ISource<IEnumerable<TypeInfo>>
 	{
 		readonly IMembers _members;