… src/ExtendedXmlSerializer/ContentModel/Xml/Parsing/GenericXmlQualifiedName.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/Parsing/NameParser.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/Parsing/QualifiedNameListParser.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/Parsing/QualifiedTypeNameFactory.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/Parsing/TypeParser.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/Parsing/TypeQualifiedNameFormatter.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/TypeExtractor.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/TypeFormatter.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/Types.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/WellKnownTypeLocator.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/XmlReader.cs
… src/ExtendedXmlSerializer/ContentModel/Xml/XmlWriter.cs
… src/ExtendedXmlSerializer/Core/Extensions.cs
… src/ExtendedXmlSerializer/Core/Sources/CachingAlteration.cs
… src/ExtendedXmlSerializer/Core/Sources/CompositeOption.cs
… src/ExtendedXmlSerializer/Core/Sources/DecoratedOption.cs
… src/ExtendedXmlSerializer/Core/Sources/DelegateAlterations.cs
… src/ExtendedXmlSerializer/Core/Sources/Items.cs
… src/ExtendedXmlSerializer/Core/Sources/ItemsBase.cs
… src/ExtendedXmlSerializer/Core/Sources/ReferenceCache.cs
… src/ExtendedXmlSerializer/Core/Sources/ReferenceCacheBase.cs
… src/ExtendedXmlSerializer/Core/Sources/ReferenceCachingAlteration.cs
… src/ExtendedXmlSerializer/Core/Specifications/IsAssignableGenericSpecification.cs
… src/ExtendedXmlSerializer/ExtendedXmlSerializer.cs
… src/ExtendedXmlSerializer/Properties/AssemblyInfo.cs
… test/ExtendedXmlSerializer.Performance.Tests/Benchmarks.cs
… test/ExtendedXmlSerializer.Performance.Tests/Program.cs
… test/ExtendedXmlSerializerTest/ContentModel/Xml/AssemblyPartitionedTypesTests.cs
… test/ExtendedXmlSerializerTest/FuturesTest.cs
… test/ExtendedXmlSerializerTest/Legacy/Resources/TestClassGenericThree.xml