Woohoo... I got some serious performance improvement that this PR introduces.  Check it out:

```
// **************************
// Benchmark: ExtendedXmlSerializerTest.DeserializationClassWithPrimitive: DefaultJob

                            Method |       Mean |    StdDev |  Gen 0 | Allocated |
---------------------------------- |----------- |---------- |------- |---------- |
   SerializationClassWithPrimitive | 74.2877 us | 0.9071 us | 7.0475 |  51.17 kB |
 DeserializationClassWithPrimitive | 97.5923 us | 0.2960 us | 5.4362 |  55.97 kB |

// **************************
// Benchmark: ExtendedXmlSerializerV2Test.DeserializationClassWithPrimitive: DefaultJob

                            Method |       Mean |    StdDev |  Gen 0 | Allocated |
---------------------------------- |----------- |---------- |------- |---------- |
   SerializationClassWithPrimitive | 33.4268 us | 0.0900 us | 4.9886 |  32.77 kB |
 DeserializationClassWithPrimitive | 44.6563 us | 0.0863 us | 3.4912 |   25.6 kB |

// **************************
// Benchmark: XmlSerializerTest.DeserializationClassWithPrimitive: DefaultJob

                            Method |       Mean |    StdDev |  Gen 0 | Allocated |
---------------------------------- |----------- |---------- |------- |---------- |
   SerializationClassWithPrimitive | 62.1980 us | 0.0671 us | 8.2031 |  56.85 kB |
 DeserializationClassWithPrimitive | 67.9054 us | 0.2292 us | 2.7832 |  27.32 kB |
```

I got this improvement by caching the primitive results (anything using `XmlConvert.*`.  Since any data that is used for these conversion is considered a constant (e.g. the string `"12.10004"` will _always_ result in a double/float/etc of `12.10004`), then this makes sense to do.  However, in case there are problems, or if someone wants to turn this off, this is supported.  We will have to add a method to the ConfigurationAPI that allows for this (maybe a `EnablePrimitiveOptimizations(bool)`).