@wojtpl2 2017-02-14
I see that in code is still XmlConvert. I don't know if I should add EnablePrimitiveOptimizations(true) or not? Or maybe do you added this?