@Mike-EEE 2017-02-14
Hey thanks. :D  Very happy with how this is turning out.  As for `EnablePrimitiveOptimizations`... I didn't add it, but was saying that we need to put one in if we want to disable it.  If you want to put it in the ConfigurationAPI I will wire it up when I am going through. 👍