@wojtpl2 2017-02-14
Ok. If we disable it (in configuration):
1. How will xml look?
2. What will you use for serialization of primitive version. I'm asking because XmlConvert generates the same xml for different regional settings (cultures).