@Mike-EEE 2017-02-14
So to be sure here, the only thing that this does it cache the results and stores it in a dictionary.  So if you ask for the `float` `3.14159`, it pulls it from the dictionary rather than asking `XmlConvert.ToSingle`.  Same with the other way as a string, and for all the primitives.  So, in theory, the XML should look the same.  However, I am not 100% on that, so that is why I think we should enable the ability to turn it off if necessary.  Also, if you have billion numbers those will all be stored in memory until the serializer is GC'd, so that is nice to prevent, too. :)